fix(protocols): move orchestra_protocols from clients to domain
clients (L6) cannot import from domain (L3). Relocate to
domain/protocols/ where GateResult/GateStatus belong naturally.

Fixes test_no_upward_imports and test_no_circular_deps_outside_l3_core.
